+++ Banka.Data/Entiteti/DeviznoPlacanje.cs
@@ -15,7 +15,7 @@ namespace Banka.Data.Entiteti
         public ulong BrojRacunaPrimaoca { get; set; }
         public int ModelPlacanja { get; set; }
         public ulong PozivNaBroj { get; set; }
-        public double Iznos { get; set; }
+        public decimal Iznos { get; set; }
         public DateTime VremePlacanja { get; set; }
         [ForeignKey("DevizniRacun")]
         public int IdDeviznogRacuna { get; set; }
