+++ Banka.DomenskaLogika/Modeli/DomenskiModeli/DeviznoPlacanjeDomenskiModel.cs
@@ -11,7 +11,7 @@ namespace Banka.DomenskaLogika.Modeli
         public ulong BrojRacunaPrimaoca { get; set; }
         public int ModelPlacanja { get; set; }
         public ulong PozivNaBroj { get; set; }
-        public double Iznos { get; set; }
+        public decimal Iznos { get; set; }
         public DateTime VremePlacanja { get; set; }
         public int IdDeviznogRacuna { get; set; }
     }
